Use HTTPS for URLs of reported tweets.
This is accomplished (at least, I hope it is) via the use of an
ed-descendent substitution construct (of which I do not at the moment
care to learn the proper Perl name), substituting the string “http:”, if
it occurs at the beginning of (what I assume to be) the reported tweet’s
URL (`$data_link`), for the string “https:”.

I do not actually know Perl (nor, having looked at it, do I particularly
want to), nor do I have any significant understanding of the Gambot
sourcebase, so this may well not work. On the other metaphorical hand,
this a quite simple change, consisting only of a single added line, and
I did test the substitution construct used, which was also used
immediately previous in the script, so this may well work.

Gambit: the file was `scripts/feed_reader/feed_twitter.pl`, not, as you
claimed at approximately 02013-07-25 02:08:28 UTC,
`scripts/StreamReader/StreamTwitter.pl`.

Also, I expect you could have done this yourself in less time than it
would take you to read this commit message and merge this commit.

Although, I have a curious suspicion that you’re not the type who will
actually bother reading this commit message, especially with your own
being so short and bereft of that essential quality that makes the act
of writing — and of reading — commit messages worthwhile, namely,
DETAIL.

Signed-off-by: 8573 <8573dd@gmail.com>