Add `memo` command.
Implement a `memo` command, to record memoranda which are to be relayed
to their designated addressees, and enable it in (what I believe to be)
the “example” and “jane” configurations.

This `memo` command is to be used thus:

    <Gambot’s nick>: memo <addressee’s nick> <memorandum text>

E.g., the following command…

    janebot: memo Gambit Lorem ipsum dolor sit amet.

…would instruct janebot to tell Gambit “Lorem ipsum dolor sit amet.”
when he’s next around (i.e., the next time he speaks in the channel in
which the memorandum was recorded).

The implementation of this `memo` command is in the following Perl
script:

    parsers/plugin_parser/plugins/memo.pm